+++ xcube_stac/accessors/sen2.py
@@ -36,8 +36,8 @@ import rioxarray
 import xarray as xr
 import xmltodict
 from xcube.core.chunk import chunk_dataset
-from xcube.core.gridmapping import GridMapping
-from xcube.core.resampling import affine_transform_dataset, reproject_dataset
+from xcube_resampling.gridmapping import GridMapping
+from xcube_resampling import resample_in_space
 from xcube.core.store import DataStoreError
 from xcube.util.jsonschema import (
     JsonArraySchema,
@@ -47,6 +47,8 @@ from xcube.util.jsonschema import (
     JsonObjectSchema,
     JsonStringSchema,
 )
+from xcube_resampling.utils import reproject_bbox
+from xcube_resampling.gridmapping import GridMapping
 
 from xcube_stac.accessor import StacArdcAccessor, StacItemAccessor
 from xcube_stac.constants import (
@@ -62,15 +64,11 @@ from xcube_stac.constants import (
 from xcube_stac.stac_extension.raster import apply_offset_scaling, get_stac_extension
 from xcube_stac.utils import (
     add_nominal_datetime,
-    get_gridmapping,
-    get_spatial_dims,
     _get_tile_size,
     merge_datasets,
     mosaic_spatial_take_first,
     normalize_crs,
-    normalize_grid_mapping,
     rename_dataset,
-    reproject_bbox,
 )
 from xcube_stac.version import version
 
@@ -295,16 +293,16 @@ class Sen2CdseStacItemAccessor(StacItemAccessor):
                 for (ds, asset) in zip(dss, assets)
             ]
         if "spatial_res" in open_params:
-            target_gm = get_gridmapping(
+            target_gm = GridMapping.regular_from_bbox(
                 bbox=assets[0].extra_fields["proj:bbox"],
-                spatial_res=open_params["spatial_res"],
+                xy_res=open_params["spatial_res"],
                 crs=assets[0].extra_fields["proj:code"],
                 tile_size=open_params.get("tile_size", TILE_SIZE),
             )
             ds = merge_datasets(dss, target_gm=target_gm)
         else:
             ds = merge_datasets(dss)
-        return normalize_grid_mapping(ds)
+        return ds
 
     def _add_sen2_angles(self, item: pystac.Item, ds: xr.Dataset) -> xr.Dataset:
         """Extract Sentinel-2 solar and viewing angle information from the granule
@@ -326,10 +324,8 @@ class Sen2CdseStacItemAccessor(StacItemAccessor):
         """
         target_gm = _get_angle_target_gm(ds)
         ds_angles = self.get_sen2_angles(item, ds)
-        ds_angles = _resample_dataset_soft(
-            ds_angles,
-            target_gm,
-            interpolation="bilinear",
+        ds_angles = resample_in_space(
+            ds_angles, target_gm=target_gm, interp_methods="bilinear"
         )
         ds = _add_angles(ds, ds_angles)
         return ds
@@ -743,10 +739,8 @@ class Sen2CdseStacArdcAccessor(Sen2CdseStacItemAccessor, StacArdcAccessor):
                 if idx not in idx_remove_dt
             ]
             ds_tile = ds_tile.assign_coords(coords=dict(time=np_datetimes_sel))
-            ds_tile = _resample_dataset_soft(
-                ds_tile,
-                target_gm,
-                interpolation="bilinear",
+            ds_tile = resample_in_space(
+                ds_tile, target_gm=target_gm, interp_methods="bilinear"
             )
             ds_tile = ds_tile.chunk(dict(time=1))
             if len(idx_remove_dt) > 0:
@@ -802,16 +796,16 @@ class Sen2PlanetaryComputerStacItemAccessor(Sen2CdseStacItemAccessor):
                 self._apply_offset_scaling(ds, item) for (ds, asset) in zip(dss, assets)
             ]
         if "spatial_res" in open_params:
-            target_gm = get_gridmapping(
+            target_gm = GridMapping.regular_from_bbox(
                 bbox=assets[0].extra_fields["proj:bbox"],
-                spatial_res=open_params["spatial_res"],
+                xy_res=open_params["spatial_res"],
                 crs=item.properties["proj:code"],
                 tile_size=open_params.get("tile_size", TILE_SIZE),
             )
             ds = merge_datasets(dss, target_gm=target_gm)
         else:
             ds = merge_datasets(dss)
-        return normalize_grid_mapping(ds)
+        return ds
 
     @staticmethod
     def _list_assets_from_item(item: pystac.Item, **open_params) -> list[pystac.Asset]:
@@ -919,7 +913,8 @@ def _get_angle_target_gm(ds_final: xr.Dataset) -> GridMapping:
         mapping used during datacube generation.
     """
     crs = pyproj.CRS.from_cf(ds_final.spatial_ref.attrs)
-    y_coord, x_coord = get_spatial_dims(ds_final)
+    gm_final = GridMapping.from_dataset(ds_final)
+    x_coord, y_coord = gm_final.xy_var_names
     if crs.is_geographic:
         y_res = 5000 / 111320
         y_center = (ds_final[x_coord][0].item() + ds_final[x_coord][-1].item()) / 2
@@ -937,7 +932,7 @@ def _get_angle_target_gm(ds_final: xr.Dataset) -> GridMapping:
     if bbox[3] < bbox[1]:
         y_min, y_max = bbox[3], bbox[1]
         bbox[1], bbox[3] = y_min, y_max
-    return get_gridmapping(bbox, (x_res, y_res), crs)
+    return GridMapping.regular_from_bbox(bbox, (x_res, y_res), crs)
 
 
 def _get_band_names_from_dataset(ds: xr.Dataset) -> list[str]:
@@ -1099,7 +1094,8 @@ def _add_angles(ds: xr.Dataset, ds_angles: xr.Dataset) -> xr.Dataset:
         - Renames spatial coordinates in `ds_angles` to avoid conflicts.
         - The angle dimension contains 'zenith' and 'azimuth'
     """
-    x_coord, y_coord = get_spatial_dims(ds_angles)
+    gm = GridMapping.from_dataset(ds_angles)
+    x_coord, y_coord = gm.xy_var_names
     ds_angles = ds_angles.rename(
         {y_coord: f"angle_{y_coord}", x_coord: f"angle_{x_coord}"}
     )
@@ -1191,52 +1187,6 @@ def _get_spatial_res(open_params: dict) -> int:
         spatial_res = int(_SEN2_SPATIAL_RES[idxs[0]])
 
     return spatial_res
-
-
-def _resample_dataset_soft(
-    ds: xr.Dataset,
-    target_gm: GridMapping,
-    interpolation: str = "nearest",
-) -> xr.Dataset:
-    """Resample a dataset to a target grid mapping, using either affine transform
-    or reprojection.
-
-    If the source and target grid mappings are close, the dataset is returned unchanged.
-    If they share the same CRS but differ spatially, an affine transformation is applied.
-    Otherwise, the dataset is reprojected to the target grid, with optional fill value
-    and interpolation.
-
-    Parameters:
-        ds: The source xarray Dataset to be resampled.
-        target_gm: The target grid mapping
-        interpolation: Interpolation method to use during reprojection
-            (see reproject_dataset function in xcube).
-
-    Returns:
-        The resampled dataset aligned with the target grid mapping.
-    """
-    source_gm = GridMapping.from_dataset(ds)
-    if source_gm.is_close(target_gm):
-        return ds
-    if target_gm.crs == source_gm.crs:
-        var_configs = {}
-        for var in ds.data_vars:
-            var_configs[var] = dict(recover_nan=True)
-        ds = affine_transform_dataset(
-            ds,
-            source_gm=source_gm,
-            target_gm=target_gm,
-            gm_name="spatial_ref",
-            var_configs=var_configs,
-        )
-    else:
-        ds = reproject_dataset(
-            ds,
-            source_gm=source_gm,
-            target_gm=target_gm,
-            interpolation=interpolation,
-        )
-    return ds
 
 
 def _create_empty_dataset(
@@ -1375,14 +1325,14 @@ def _merge_utm_zones(list_ds_utm: list[xr.Dataset], **open_params) -> xr.Dataset
             ds.x[1] - ds.x[0] != spatial_res[0]
             or abs(ds.y[1] - ds.y[0]) != spatial_res[1]
         ):
-            target_gm = get_gridmapping(
+            target_gm = GridMapping.regular_from_bbox(
                 open_params["bbox"],
                 open_params["spatial_res"],
                 open_params["crs"],
                 tile_size=open_params.get("tile_size", TILE_SIZE),
             )
     else:
-        target_gm = get_gridmapping(
+        target_gm = GridMapping.regular_from_bbox(
             open_params["bbox"],
             open_params["spatial_res"],
             open_params["crs"],
@@ -1391,7 +1341,9 @@ def _merge_utm_zones(list_ds_utm: list[xr.Dataset], **open_params) -> xr.Dataset
 
     resampled_list_ds = []
     for ds in list_ds_utm:
-        resampled_list_ds.append(_resample_dataset_soft(ds, target_gm))
+        resampled_list_ds.append(
+            resample_in_space(ds, target_gm=target_gm, prevent_nan_propagations=True)
+        )
     return mosaic_spatial_take_first(resampled_list_ds)